@@ -94,7 +94,7 @@ build-backend = "poetry.core.masonry.api"
 # "tests",
 # "neps_examples",
 [tool.ruff]
-target-version = "py38"
+target-version = "py310"
 output-format = "full"
 line-length = 90
 src = ["neps"]
