@@ -3,7 +3,8 @@
 from __future__ import annotations
 
 import os
-from typing import Callable, TypeVar
+from collections.abc import Callable
+from typing import TypeVar
 
 T = TypeVar("T")
 V = TypeVar("V")
