@@ -23,7 +23,8 @@ set or empty, in which case it is `None`.
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
-from typing import Any, ClassVar, Generic, Mapping, TypeVar, runtime_checkable
+from collections.abc import Mapping
+from typing import Any, ClassVar, Generic, TypeVar, runtime_checkable
 from typing_extensions import Protocol, Self
 
 ValueT = TypeVar("ValueT")
