@@ -7,18 +7,15 @@ import logging
 import os
 import shutil
 import time
+from collections.abc import Callable, Iterable, Iterator, Mapping
 from contextlib import contextmanager
 from dataclasses import dataclass
 from pathlib import Path
 from typing import (
     TYPE_CHECKING,
     Any,
-    Callable,
     Generic,
-    Iterable,
-    Iterator,
     Literal,
-    Mapping,
     TypeVar,
 )
 
