@@ -3,9 +3,10 @@
 from __future__ import annotations
 
 import dataclasses
+from collections.abc import Iterable, Mapping
 from enum import Enum
 from pathlib import Path
-from typing import Any, Iterable, Mapping
+from typing import Any
 
 import yaml
 
